ext/dfbvideosink: set the CbCr plane address on semi-planar YUV buffers
is_ycbcr_planar() returns true only if a color format is planar.
It has been replaced with is_ycbcr() because the address of CbCr
plane also must be set in semi planar formats. In order to omit packed
YUV, a test to determind whether or not the format is UYVY was added.