ext/dfbvideosink: convert interlaced video fields into a frame
This makes the renderer check the two capabilities 'interlaced' and
'field-layout' attached with input video frame and deinterlaces it
if needed. Unfortunately this could only support
'field-layout=sequential' which means an image contain both fields,
the top field lines are stored first in a buffer, immediately
followed by the bottom field lines.