ext/dfbvideosink: make sure that the last line rendered is from the top field when an output image has odd height
When an output image had odd height, there was a one-line gap at the
bottom of the image. It was due to an extra line of the bottom field
instead of the top field being rendered by mistake. This incorrect behavior
created a gap above the last line after the progressive conversion.
This patch fixes it to render the last line of the top field when
the height is odd.