ext/dfbvideosink: clear MERAM whenever non 16-byte aligned address or width specified
MERAM must be aligned on 16-byte boundaries. If non-16 byte aligned
address or width is specified for output destination, the MERAM may
write image data with unnessesary data that exists from the previous
16-byte boundary to the address.

This fills out the MERAM with background color data to prevent to
show such the noise in display.