h264parse: flush the last nal in the GstAdapter when a EOS event is caught
When an H.264 bytestream is sent by a previous plugin, a last nal
packet remains in the GstAdapter and is cleared after playback ends.
h264parse looks at a next nal startcode in order to detect nal boundaries,
and so it can't find the end of the last nal.
This change pushes a nal start code to the adapter when an EOS event is caught.

This patch is released by Pioneer.