ext/dfbvideosink: calculate Y plane stride when the color format is planar or semi planar
src_pitch is determined by dividing the buffer size by the image height, which
is correct when a packed format is specified. However, When the planar or
the semi planar fomats are specified, src_picth (for the Y plane)
must be adjusted to account for the CbCr plane that is present in the same
buffer.