ext/dfbvideosink: fix condition of the MERAM clearing to treat auto-resizing with 'keep-aspect-ratio=true'
The process of MERAM clearing is only necessary if either the start
or end of a line that the VIO writes to is not a 16-byte aligned
address.
However, if the 'keep-aspect-ratio' property is true, the size and
position of the final output image may not match the value set by
the 'window-*' properties, causing garbage to be left uncleared in
the MERAM.
This activates the MERAM clearing regardless of the 'window-*'
properties if 'keep-aspect-ratio' is true.